feat: `MinePlacer` 분리
