+++ src/main/kotlin/minesweeper/Board.kt
@@ -1,7 +1,5 @@
 package minesweeper
 
-import kotlin.random.Random
-
 class Board(
     private val dimensions: Dimensions,
     minePositions: Set<Position>,
@@ -35,26 +33,5 @@ class Board(
 
     companion object {
         const val MIN_MINE_COUNT = 1
-    }
-}
-
-class MinePlacer(
-    private val dimensions: Dimensions,
-    private val mineCount: Int,
-) {
-    init {
-        require(mineCount in 1 until dimensions.totalCells) {
-            "Mine count must be less than total cells and greater than 0."
-        }
-    }
-
-    fun placeMines(): Set<Position> {
-        val positions = mutableSetOf<Position>()
-        while (positions.size < mineCount) {
-            val x = Random.nextInt(dimensions.width)
-            val y = Random.nextInt(dimensions.height)
-            positions.add(Position(x, y))
-        }
-        return positions
     }
 }
